Long cast

+++ bookkeeper-server/src/test/java/org/apache/bookkeeper/bookie/storage/ldb/WriteCacheTest.java
@@ -86,7 +86,7 @@ public class WriteCacheTest {
             }
         }
         for(int c = 0; c <10; c ++) {
-        	long d = 0.0 + c;
+        	long d = (long)0 + (long)c;
              assertEquals(entry, cache.get(1.0,1.0));
              assertEquals((cache.get(d,d)),entry);
              cache.deleteLedger(d);
